layout/base/tests/test_bug404209.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=404209
-->
<head>
  <title>Test for Bug 404209</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style type="text/css">
   div::first-letter { color: green; }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=404209">Mozilla Bug 404209</a>
<table id="table" dir="rtl"><div><span id="v"><span><tfoot></tfoot>abcd</span></span></div></table>
<pre id="test">
<script class="testbody" type="text/javascript">
function boom1()
{
  document.getElementById("table").style.borderRight = "1px solid magenta";
  setTimeout(boom2, 400);
}

function boom2()
{
  var v = document.getElementById("v"); 
  var newTD = document.createElementNS("http://www.w3.org/1999/xhtml", "td");
  newTD.setAttribute("width", "13%");
  v.insertBefore(newTD, v.firstChild);
  setTimeout(lastTest, 400);
}

function lastTest()
{
/** Test for Bug 404209 **/
  ok(true, "Should not crash");
  SimpleTest.finish();
}

  SimpleTest.requestFlakyTimeout("untriaged");
  setTimeout(boom1, 400);
  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
